chore: restructure feat: info about the cube shows up BUG: Text in the wrong color
